Move OrderBy and Direction to segment::data_types::order_by
