+++ lib/collection/src/shards/proxy_shard.rs
@@ -5,8 +5,9 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use async_trait::async_trait;
+use segment::data_types::order_by::OrderBy;
 use segment::types::{
-    ExtendedPointId, Filter, OrderBy, PointIdType, ScoredPoint, WithPayload, WithPayloadInterface,
+    ExtendedPointId, Filter, PointIdType, ScoredPoint, WithPayload, WithPayloadInterface,
     WithVector,
 };
 use tokio::runtime::Handle;
