+++ lib/segment/src/types.rs
@@ -1977,67 +1977,6 @@ pub struct WithPayload {
     pub payload_selector: Option<PayloadSelector>,
 }
 
-#[derive(Debug, Deserialize, Serialize, JsonSchema, Copy, Clone, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum Direction {
-    #[default]
-    Asc,
-    Desc,
-}
-
-impl Direction {
-    pub fn as_range(&self, from: f64) -> Range {
-        match self {
-            Direction::Asc => Range {
-                gte: Some(from),
-                ..Default::default()
-            },
-            Direction::Desc => Range {
-                lte: Some(from),
-                ..Default::default()
-            },
-        }
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize, JsonSchema, Validate, Clone, Default)]
-#[serde(deny_unknown_fields)]
-#[serde(rename_all = "snake_case")]
-pub struct OrderBy {
-    /// Payload key to order by
-    pub key: String,
-
-    /// Direction of ordering: `asc` or `desc`. Default is ascending.
-    pub direction: Option<Direction>,
-
-    /// Used internally (not exposed to the user in scroll) to know which payload value to start scrolling from,
-    /// we get this offset value by fetching the offset id in the request
-    #[serde(skip)]
-    pub value_offset: Option<f64>,
-}
-
-impl OrderBy {
-    pub fn as_range(&self) -> Range {
-        match self.value_offset {
-            Some(offset) => self.direction.unwrap().as_range(offset),
-            None => Range {
-                ..Default::default()
-            },
-        }
-    }
-
-    pub fn direction(&self) -> Direction {
-        self.direction.unwrap_or_default()
-    }
-
-    pub fn value_offset(&self) -> f64 {
-        self.value_offset.unwrap_or_else(|| match self.direction() {
-            Direction::Asc => f64::NEG_INFINITY,
-            Direction::Desc => f64::INFINITY,
-        })
-    }
-}
-
 #[derive(Debug, Deserialize, Serialize, JsonSchema, Validate, Clone, PartialEq, Default)]
 #[serde(deny_unknown_fields, rename_all = "snake_case")]
 pub struct Filter {
